refactoring: Models/AccountViewModels
