+++ ASC.Web/Models/AccountViewModels/ForgotPasswordViewModel.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ASC.Web.Models
+namespace ASC.Web.Models.AccountViewModels
 {
     public class ForgotPasswordViewModel
     {
