+++ ASC.Web/Models/AccountViewModels/SendCodeViewModel.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ASC.Web.Models
+namespace ASC.Web.Models.AccountViewModels
 {
     public class SendCodeViewModel
     {
