+++ ASC.Web/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model ForgotPasswordViewModel
+﻿@model ASC.Web.Models.AccountViewModels.ForgotPasswordViewModel
 @{
     ViewData["Title"] = "Forgot your password?";
 }
