+++ ASC.Web/Views/Account/Login.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@model LoginViewModel
+@model ASC.Web.Models.AccountViewModels.LoginViewModel
 @inject SignInManager<ApplicationUser> SignInManager
 @{
     ViewData["Title"] = "Log in";
