+++ ASC.Web/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model ResetPasswordViewModel
+﻿@model ASC.Web.Models.AccountViewModels.ResetPasswordViewModel
 @{
     ViewData["Title"] = "Reset password";
     Layout = "_Layout";
